pyproject.toml
[project]
name = "hs-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "django>=6.0.4",
    "django-cors-headers>=4.9.0",
    "django-ninja>=1.6.2",
    "google-genai>=1.72.0",
    "gunicorn>=25.3.0",
    "httpx>=0.28.1",
    "markdownify>=1.2.2",
    "pyjwt>=2.10.1",
    "pytest-django>=4.12.0",
    "pytest>=9.0.3",
    "ruff>=0.15.10",
    "whitenoise>=6.9.0",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "backend.settings"
python_files = ["test_*.py", "*_test.py"]
testpaths = ["auth/tests", "pagechecker/tests"]
